fix(cpp): resolve wave-3 mechanical clang-tidy findings
Default-initialize locals, replace unsafe static downcasts with
dynamic_cast, fix narrowing conversions, and use int64_t instead of
long long in targeted catalog, semantic, storage, and handler files.

+++ backend/catalog/info_schema_table.cc
@@ -317,7 +317,7 @@ InfoSchemaTable::CreateEvaluatorTableIterator(
   if (column_idxs.empty()) {
     idxs.resize(row_schema_.columns.size());
     for (size_t i = 0; i < row_schema_.columns.size(); ++i)
-      idxs[i] = i;
+      idxs[i] = static_cast<int>(i);
   } else {
     idxs.assign(column_idxs.begin(), column_idxs.end());
   }
